rough version of "fix for chunked-compression with hi-lo-byte-unpacking" (do hi-lo-byte-unpacking per chunk, not per bitmap)

+++ Src/libCZI/decoder_chunkedcompression.cpp
@@ -291,12 +291,15 @@ std::shared_ptr<libCZI::IBitmapData> CChunkedCompressionDecoder::Decode(const vo
             throw runtime_error("Decompressed chunk size does not match the uncompressed size declared in the chunked-compression header.");
         }
 
+        // Each chunk is a self-contained hi-lo packed flat buffer (no row-stride concerns).
+        // The encoder produces the hi-lo layout per-chunk, so we unpack it the same way:
+        // treat the whole chunk as a single row of (decompressed_size / 2) 16-bit pixels.
         LoHiBytePackUnpack::LoHiBytePackStrided(
             staging_buffer.data(),
             decompressed_size,
-            decompressed_size / 2,
+            static_cast<uint32_t>(decompressed_size / 2),
             1,
-            decompressed_size,
+            static_cast<uint32_t>(decompressed_size),
             static_cast<uint8_t*>(destination) + destination_offset);
 
         destination_offset += chunk.uncompressedSize;
